requirements.txt
asgiref==3.8.1
colorama==0.4.6
distlib==0.3.9
Django==5.2
filelock==3.17.0
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
platformdirs==4.3.6
pyzbar==0.1.9
qrcode==8.2
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.29.2
